requirements.txt
pybluez[ble]
gatt
pyyaml 
pytests
paho-mqtt